\name{rcpparrowc-package}
\alias{rcpparrowc-package}
\alias{rcpparrowc}
\docType{package}
\title{\packageTitle{rcpparrowc}}
\description{\packageDescription{rcpparrowc}}
\section{Package Content}{\packageIndices{rcpparrowc}}
\author{\packageAuthor{rcpparrowc}}
\section{Maintainer}{\packageMaintainer{rcpparrowc}}
\keyword{package}
