\name{rcppnanoarrow-package}
\alias{rcppnanoarrow-package}
\alias{rcppnanoarrow}
\docType{package}
\title{\packageTitle{rcppnanoarrow}}
\description{\packageDescription{rcppnanoarrow}}
\section{Package Content}{\packageIndices{rcppnanoarrow}}
\author{\packageAuthor{rcppnanoarrow}}
\section{Maintainer}{\packageMaintainer{rcppnanoarrow}}
\keyword{package}
